Updates to New EOI component.
